backend/requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
web3
python-dotenv
